# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  admin.CreateNamedRequest:
    properties:
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  admin.NamedResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  admin.UpdateNamedRequest:
    properties:
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  response.PageData:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Response:
    properties:
      cause:
        type: string
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: 小橘TV - 影视站系统
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 小橘TV
  version: "1.0"
paths:
  /api/admin/v1/actors:
    get:
      consumes:
      - application/json
      description: 分页获取演员列表
      parameters:
      - in: query
        name: keyword
        type: string
      - in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - in: query
        maximum: 1000000
        minimum: 1
        name: page
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageData'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/admin.NamedResponse'
                        type: array
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: 演员列表
      tags:
      - 管理端｜元数据管理
    post:
      consumes:
      - application/json
      description: 创建一个新的演员
      parameters:
      - description: 演员参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.CreateNamedRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/admin.NamedResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 新建演员
      tags:
      - 管理端｜元数据管理
  /api/admin/v1/actors/{id}:
    delete:
      description: 删除指定演员
      parameters:
      - description: 演员ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除演员
      tags:
      - 管理端｜元数据管理
    put:
      consumes:
      - application/json
      description: 更新指定演员信息
      parameters:
      - description: 演员ID
        in: path
        name: id
        required: true
        type: integer
      - description: 演员参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.UpdateNamedRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/admin.NamedResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 更新演员
      tags:
      - 管理端｜元数据管理
...
  /api/admin/v1/directors:
    get:
      consumes:
      - application/json
      description: 分页获取导演列表
      parameters:
      - in: query
        name: keyword
        type: string
      - in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - in: query
        maximum: 1000000
        minimum: 1
        name: page
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageData'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/admin.NamedResponse'
                        type: array
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: 导演列表
      tags:
      - 管理端｜元数据管理
    post:
      consumes:
      - application/json
      description: 创建一个新的导演
      parameters:
      - description: 导演参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.CreateNamedRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/admin.NamedResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 新建导演
      tags:
      - 管理端｜元数据管理
  /api/admin/v1/directors/{id}:
    delete:
      description: 删除指定导演
      parameters:
      - description: 导演ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除导演
      tags:
      - 管理端｜元数据管理
    put:
      consumes:
      - application/json
      description: 更新指定导演信息
      parameters:
      - description: 导演ID
        in: path
        name: id
        required: true
        type: integer
      - description: 导演参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.UpdateNamedRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/admin.NamedResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 更新导演
      tags:
      - 管理端｜元数据管理
...
  /api/admin/v1/tags:
    get:
      consumes:
      - application/json
      description: 分页获取标签列表
      parameters:
      - in: query
        name: keyword
        type: string
      - in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - in: query
        maximum: 1000000
        minimum: 1
        name: page
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageData'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/admin.NamedResponse'
                        type: array
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: 标签列表
      tags:
      - 管理端｜元数据管理
    post:
      consumes:
      - application/json
      description: 创建一个新的标签
      parameters:
      - description: 标签参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.CreateNamedRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/admin.NamedResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 新建标签
      tags:
      - 管理端｜元数据管理
  /api/admin/v1/tags/{id}:
    delete:
      description: 删除指定标签
      parameters:
      - description: 标签ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除标签
      tags:
      - 管理端｜元数据管理
    put:
      consumes:
      - application/json
      description: 更新指定标签信息
      parameters:
      - description: 标签ID
        in: path
        name: id
        required: true
        type: integer
      - description: 标签参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.UpdateNamedRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/admin.NamedResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 更新标签
      tags:
      - 管理端｜元数据管理
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"